Support ext-foreign-toplevel-list, refactor
This adds support for ext-foreign-toplevel-list protocol, which will replace
the zwlr-foreign-toplevel-management protocol eventually. That one is still
supported as well though for now, but it will be deprecated eventually.

Also refactor in favour of new protocol: This means the removal of the output
and state information. The new protocol does not provide these, instead it is
intended as a basis for compositor-specific extensions. Such extensions can be
used to communicate toplevel state. Should any extensions pop up and gain
popularity, lswt will support them, which should be trivial.

With the removal of output information, there is no longer any relational
information, meaning the dot output format had to be removed as well for now.